+++ plugins/org.projectforge.plugins.merlin/src/main/kotlin/org/projectforge/plugins/merlin/rest/MerlinAttachmentsActionListener.kt
@@ -50,7 +50,7 @@ class MerlinAttachmentsActionListener(
   /**
    * Allows only upload of Word and Excel documents.
    */
-  override fun onBeforeUpload(fileInfo: FileInfo, obj: ExtendedBaseDO<Int>): ResponseEntity<*>? {
+  override fun onBeforeUpload(fileInfo: FileInfo, obj: ExtendedBaseDO<Long>): ResponseEntity<*>? {
     return if (fileInfo.fileExtension != "docx" && fileInfo.fileExtension != "xlsx") {
       ResponseEntity.ok().body(
         UIToast.createToast(
@@ -68,7 +68,7 @@ class MerlinAttachmentsActionListener(
    */
   override fun afterUpload(
     attachment: Attachment,
-    obj: ExtendedBaseDO<Int>,
+    obj: ExtendedBaseDO<Long>,
     jcrPath: String,
     attachmentsAccessChecker: AttachmentsAccessChecker,
     listId: String?
@@ -97,7 +97,7 @@ class MerlinAttachmentsActionListener(
 
   override fun afterModification(
     attachment: Attachment,
-    obj: ExtendedBaseDO<Int>,
+    obj: ExtendedBaseDO<Long>,
     jcrPath: String,
     attachmentsAccessChecker: AttachmentsAccessChecker,
     listId: String?
@@ -107,7 +107,7 @@ class MerlinAttachmentsActionListener(
   }
 
   private fun createResponseEntity(
-    obj: ExtendedBaseDO<Int>,
+    obj: ExtendedBaseDO<Long>,
     list: List<Attachment>?,
     targetType: TargetType
   ): ResponseEntity<*> {
@@ -127,7 +127,7 @@ class MerlinAttachmentsActionListener(
   }
 
   override fun createAttachmentLayout(
-    id: Int,
+    id: Long,
     category: String,
     fileId: String,
     listId: String?,
