+++ projectforge-business/src/main/java/org/projectforge/business/fibu/api/EmployeeSalaryService.java
@@ -36,6 +36,6 @@ public interface EmployeeSalaryService
 
   void saveOrUpdate(EmployeeSalaryDO employeeSalaryDO);
 
-  EmployeeSalaryDO selectByPk(Integer id);
+  EmployeeSalaryDO selectByPk(Long id);
 
 }
