+++ projectforge-business/src/main/java/org/projectforge/business/fibu/kost/KostZuweisungDao.java
@@ -83,9 +83,9 @@ public class KostZuweisungDao extends BaseDao<KostZuweisungDO>
   /**
    * @param kostZuweisung
    * @param kost1Id If null, then kost1 will be set to null;
-   * @see BaseDao#getOrLoad(Integer)
+   * @see BaseDao#getOrLoad(Long)
    */
-  public void setKost1(final KostZuweisungDO kostZuweisung, Integer kost1Id)
+  public void setKost1(final KostZuweisungDO kostZuweisung, Long kost1Id)
   {
     Kost1DO kost1 = kost1Dao.getOrLoad(kost1Id);
     kostZuweisung.setKost1(kost1);
@@ -93,10 +93,9 @@ public class KostZuweisungDao extends BaseDao<KostZuweisungDO>
 
   /**
    * @param kostZuweisung
-   * @param kost1Id If null, then kost2 will be set to null;
-   * @see BaseDao#getOrLoad(Integer)
+   * @see BaseDao#getOrLoad(Long)
    */
-  public void setKost2(final KostZuweisungDO kostZuweisung, Integer kost2Id)
+  public void setKost2(final KostZuweisungDO kostZuweisung, Long kost2Id)
   {
     Kost2DO kost2 = kost2Dao.getOrLoad(kost2Id);
     kostZuweisung.setKost2(kost2);
